Switch to a better API from PostNord.
This API is more officially supported, and can find parcels the old one couldn't.
It requires a "consumer ID", which due to their rules must be kept secret, so it's not included in the open source code. Requesting a consumer ID for a fork is, however, very simple (there's a link in ParcelJsonParser.java).

The new API has so far not been tested thoroughly, but the current code works without issues for my saved collection of mixed parcel ID:s. Of course, most of those parces have already been delivered...

This update will wipe the events table in the database, since the two API:s differ a lot in representation there (would result i duplicate events otherwise).